+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fix an issue introduced in 6.8.0 where audio could be desynced on video seeking.
+- Do better validation of the retail PTR log path.
 
 ## [6.8.0] - 2025-02-21
 ### Added